pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[project]
name = "usbman"
dynamic = ["version"]
authors = [
  { name="Sebastien Riou", email="" },
]
description = "Controller for USB managed hubs"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
  "pysatl>=1.2.6",
  "pyserial"
]
[project.optional-dependencies]
doc = [
  "sphinx",
]
[project.urls]
"Homepage" = "https://github.com/sebastien-riou/usbman"
"Bug Tracker" = "https://github.com/sebastien-riou/usbman/issues"


[tool.black]
target-version = ["py310"]
line-length = 120
skip-string-normalization = true

[tool.ruff]
target-version = "py310"
line-length = 120
lint.select = [
  "A",
  "B",
  "C",
  "E",
  "F",
  "FBT",
  "I",
  "N",
  "Q",
  "RUF",
  "S",
  "T",
  "UP",
  "W",
  "YTT",
]
lint.ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Ignore McCabe complexity
  "C901",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Allow the class names to not use CapWords
  "N801",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  #below are those that we would like to deal with in the long term
  "UP031", # replace % by fstring
  "RUF013", #PEP 484 prohibits implicit `Optional`
  "S101",#S101 Use of `assert` detected
  "E501", #Line too long (130 > 120)
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
]
lint.unfixable = [
  # Don't touch unused imports
  "F401",
]
exclude = ["docs","test","dist"]

[tool.ruff.lint.isort]
known-first-party = ["hdrbg"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
# Tests can use relative imports and assertions and print
"test/**/*" = ["TID252", "S101", "T201"]
# CLI can print
"usbman/clicom.py" = ["T201"]
"usbman/__main__.py" = ["T201","I001"]

[tool.mypy]
disallow_untyped_defs = false
follow_imports = "normal"
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_codes = true
warn_no_return = false
warn_unused_ignores = true

[tool.coverage.paths]
usbman = ["usbman", "*/usbman/*"]


[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]